defined format of activities and provided examples
In the resource folder there are example of how they information for the activities will be created. Each folder has a number of .yaml files that will eventually end up in the database. yaml is a very simple file format and should be self-explanatory. Simply copy and paste the example to make a new activity.  modules.yaml defines the list of modules.

These .yaml activities will be eventually inserted into the database using some Python helper.